aklog: Fix some format warnings
Fix some format warnings (size_t vs int) which only appear when we're
building with Heimdal.

Reviewed-on: http://gerrit.openafs.org/2740
Tested-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit 5bc4b4956509256e8cf13bd04e4615ee2cb31c65)

Reviewed-on: http://gerrit.openafs.org/3986
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/aklog/aklog.c
@@ -1985,8 +1985,8 @@ get_credv5_akimpersonate(krb5_context context,
 
         if(buf_len != buf_size) {
             afs_com_err(progname, code,
-		    "%d != %d while encoding ticket (internal ASN.1 encoder error",
-		    buf_len, buf_size);
+		    "%u != %u while encoding ticket (internal ASN.1 encoder error",
+		    (unsigned int)buf_len, (unsigned int)buf_size);
             goto cleanup;
         }
         what = "krb5_crypto_init";
